% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDfunctions.R
\name{mkGeneDistPlots}
\alias{mkGeneDistPlots}
\title{Create gene expression distribution plots}
\usage{
mkGeneDistPlots(GeneExpressionMat)
}
\arguments{
\item{GeneExpressionMat}{A data frame or matrix where rows are genes and columns are samples.}
}
\value{
A list of ggplot objects, one for each sample.
}
\description{
Creates histogram plots of gene expression distribution for each sample.
}
\details{
This function creates distribution plots for gene expression data. It removes
zero-count genes, transforms data to log2 scale, and adds reference lines for
median expression and defined thresholds (LowP50 and HighP50).

Note: The variables LowP50 and HighP50 need to be defined before calling this function.
}
\examples{
\dontrun{
# Define threshold values
LowP50 <- 2
HighP50 <- 8

# Create distribution plots
plots <- mkGeneDistPlots(expression_matrix)

# Save plots to a directory
dir.create("expression_plots", showWarnings = FALSE)
for (i in 1:length(plots)) {
  ggsave(paste0("expression_plots/", names(plots)[i], ".pdf"),
         plots[[i]], width = 8, height = 6)
}
}

}
